Fixing POST content type on InstaCount
With the switch to Echo on the back end, it is much more particular about the
content type of the request it receives. Previous code put JSON in the request
body but would set the content type to be form encoded, which is wrong.
This change sends the proper application/json content type.