Unique Conditions and Medications
In patient summary unique the conditions and meds by their display names to minimize duplicates displayed to users

Why does Ember not have a uniqBy function on Array? It does, but it wasn't added until, I don't know because accessing old documentation is apparently not a desired use case of the website. (includes LTS release documentation)